subiendo mas archivos de imagenes

+++ resources/views/radicados/show.blade.php
@@ -315,30 +315,39 @@
     	                        </td>
     						</tr>
     					@endif
-    					@if($radicado->adjunto)
+    					@if(($radicado->adjunto) || ($radicado->adjunto_1) || ($radicado->adjunto_2) || ($radicado->adjunto_3) || ($radicado->adjunto_4))
     					    <tr id="tr_adjunto">
     							<th>Archivos Adjunto</th>
     							<td>
-                                    {{-- Archivo 1 --}}
-    								<a href="{{asset('adjuntos/documentos/'.$radicado->adjunto)}}" target="_blank" class="btn btn-outline-success btn-sm btn-icons" style="border-radius: 50%;" title="Ver Adjunto"><i class="fas fa-eye"></i>
-    								<a href="javascript:eliminar('{{$radicado->id}}')" class="btn btn-outline-danger btn-sm btn-icons ml-1" style="border-radius: 50%;" title="Eliminar Adjunto"><i class="fas fa-times"></i></a>
+                                    @if ($radicado->adjunto)
+                                        {{-- Archivo 1 --}}
+    								    <a href="{{asset('adjuntos/documentos/'.$radicado->adjunto)}}" target="_blank" class="btn btn-outline-success btn-sm btn-icons" style="border-radius: 50%;" title="Ver Adjunto"><i class="fas fa-eye"></i>
+                                        <a href="javascript:eliminar('{{$radicado->id}}')" class="btn btn-outline-danger btn-sm btn-icons ml-1" style="border-radius: 50%;" title="Eliminar Adjunto"><i class="fas fa-times"></i></a>
+                                    @endif
+                                    @if ($radicado->adjunto_1)
                                     {{-- Archivo 2 --}}
 
                                     <a href="{{asset('adjuntos/documentos/'.$radicado->adjunto_1)}}" target="_blank" class="btn btn-outline-success btn-sm btn-icons" style="border-radius: 50%;" title="Ver Adjunto"><i class="fas fa-eye"></i>
                                     <a href="javascript:eliminar('{{$radicado->id}}')" class="btn btn-outline-danger btn-sm btn-icons ml-1" style="border-radius: 50%;" title="Eliminar Adjunto"><i class="fas fa-times"></i></a>
                                     {{-- Archivo 3 --}}
-
+                                    @endif
+                                    @if ($radicado->adjunto_2)
                                     <a href="{{asset('adjuntos/documentos/'.$radicado->adjunto_2)}}" target="_blank" class="btn btn-outline-success btn-sm btn-icons" style="border-radius: 50%;" title="Ver Adjunto"><i class="fas fa-eye"></i>
                                     <a href="javascript:eliminar('{{$radicado->id}}')" class="btn btn-outline-danger btn-sm btn-icons ml-1" style="border-radius: 50%;" title="Eliminar Adjunto"><i class="fas fa-times"></i></a>
                                     {{-- Archivo 4 --}}
-
+                                    @endif
+                                    @if ($radicado->adjunto_4)
                                     <a href="{{asset('adjuntos/documentos/'.$radicado->adjunto_3)}}" target="_blank" class="btn btn-outline-success btn-sm btn-icons" style="border-radius: 50%;" title="Ver Adjunto"><i class="fas fa-eye"></i>
                                     <a href="javascript:eliminar('{{$radicado->id}}')" class="btn btn-outline-danger btn-sm btn-icons ml-1" style="border-radius: 50%;" title="Eliminar Adjunto"><i class="fas fa-times"></i></a>
+                                    @endif
                                     {{-- Archivo 5 --}}
-
+                                    @if ($radicado->adjunto_4)
                                     <a href="{{asset('adjuntos/documentos/'.$radicado->adjunto_4)}}" target="_blank" class="btn btn-outline-success btn-sm btn-icons" style="border-radius: 50%;" title="Ver Adjunto"><i class="fas fa-eye"></i>
                                     <a href="javascript:eliminar('{{$radicado->id}}')" class="btn btn-outline-danger btn-sm btn-icons ml-1" style="border-radius: 50%;" title="Eliminar Adjunto"><i class="fas fa-times"></i></a>
-    	                        </td>
+                                    @endif
+
+
+                                </td>
 
     						</tr>
     					@endif
